feat:添加优惠卷功能

+++ model/src/main/java/com/yzx/model/product/vo/CategoryVo.java
@@ -17,7 +17,7 @@ public class CategoryVo implements Serializable {
     private static final long serialVersionUID = 1L;
     private Long catId;
     private String name;
-    private Integer parentCid;
+    private Long parentCid;
     private Integer catLevel;
     private Integer showStatus;
     private String icon;
